requirements.txt
Flask
Pillow
pytesseract
boto3
opencv-python-headless